Split of routing into seperate module, removed settings component
