+++ src/app/app-routing.module.ts
@@ -2,36 +2,46 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { LoginComponent } from './login/login.component';
 import { HomeComponent } from './home/home.component';
-import { SettingsComponent } from './settings/settings.component';
 import { AuthGuard } from './auth-guard.service';
 
-const appRoutes: Routes = [
-  {
-    path: 'login',
-    component: LoginComponent
-  },
+import { SetupGuard } from './setup-guard.service';
+import { SetupComponent } from './setup/setup.component';
+
+const appRoutes = [
   {
     path: '',
-    canActivate: [AuthGuard],
+    canActivate: [SetupGuard],
     children: [
       {
-        path: 'home',
-        component: HomeComponent
+        path: 'login',
+        component: LoginComponent,
+        canActivate: [AuthGuard]
       },
       {
-        path: 'settings',
-        component: SettingsComponent
+        path: '',
+        canActivate: [AuthGuard],
+        children: [
+          {
+            path: 'home',
+            component: HomeComponent
+          }
+        ]
       }
     ]
+  },
+  {
+    path: 'setup',
+    canActivate: [SetupGuard],
+    component: SetupComponent
   }
 ];
 
 @NgModule({
   imports: [
-    RouterModule.forChild(appRoutes)
+    RouterModule.forRoot(appRoutes)
   ],
   exports: [
     RouterModule
   ]
 })
-export class AppRoutingModule {}
+export class AppRoutingModule { }
