--- src/app/settings/settings.component.html
@@ -1,20 +1,0 @@
-<h1>Settings</h1>
-<div>
-  <form [formGroup]="settingsForm" novalidate (ngSubmit)="saveSettings(settingsForm.value, settingsForm.valid)">
-    <md-input-container>
-      <input mdInput type="text" placeholder="Plex Host" formControlName="plexhost">
-    </md-input-container>
-
-    <md-input-container>
-      <input mdInput type="text" placeholder="Plex Port" formControlName="plexport">
-    </md-input-container>
-
-    <md-input-container>
-      <input mdInput type="text" placeholder="Username" formControlName="username">
-    </md-input-container>
-
-    <md-input-container>
-      <input mdInput type="password" placeholder="Password" formControlName="password">
-    </md-input-container>
-  </form>
-</div>
